split utils tests into smaller files as well
